requirements.txt
gradio
PyMuPDF
faiss-cpu
pyyaml
psutil
openai
langchain
langchain-openai